feat: added load members action (#790)
