+++ app/Actions/Models/Wiki/ApiAction.php
@@ -5,11 +5,6 @@ declare(strict_types=1);
 namespace App\Actions\Models\Wiki;
 
 use App\Enums\Models\Wiki\ResourceSite;
-use App\Models\BaseModel;
-use App\Models\Wiki\Anime;
-use App\Models\Wiki\ExternalResource;
-use App\Models\Wiki\Studio;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Support\Arr;
