+++ app/Filament/Resources/Wiki/Anime.php
@@ -10,7 +10,6 @@ use App\Enums\Models\Wiki\ResourceSite;
 use App\Filament\Actions\Discord\DiscordThreadAction;
 use App\Filament\Actions\Models\Wiki\Anime\AttachAnimeResourceAction;
 use App\Filament\Actions\Models\Wiki\Anime\BackfillAnimeAction;
-use App\Filament\Actions\Models\Wiki\AttachImageAction;
 use App\Filament\Components\Columns\TextColumn;
 use App\Filament\Components\Fields\Select;
 use App\Filament\Components\Fields\Slug;
@@ -367,8 +366,6 @@ class Anime extends BaseResource
                     DiscordThreadAction::make('discord-thread'),
 
                     BackfillAnimeAction::make('backfill-anime'),
-
-                    AttachImageAction::make('attach-anime-image'),
 
                     AttachAnimeResourceAction::make('attach-anime-resource'),
