+++ app/Filament/Resources/Wiki/Artist/Pages/EditArtist.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\Filament\Resources\Wiki\Artist\Pages;
 
 use App\Filament\HeaderActions\Models\Wiki\Artist\AttachArtistResourceHeaderAction;
-use App\Filament\HeaderActions\Models\Wiki\AttachImageHeaderAction;
 use App\Filament\Resources\Base\BaseEditResource;
 use App\Filament\Resources\Wiki\Artist;
 use Filament\Actions\ActionGroup;
@@ -30,8 +29,6 @@ class EditArtist extends BaseEditResource
             parent::getHeaderActions(),
             [
                 ActionGroup::make([
-                    AttachImageHeaderAction::make('attach-artist-image'),
-
                     AttachArtistResourceHeaderAction::make('attach-artist-resource'),
                 ])
             ],
