+++ app/Models/Admin/Report.php
@@ -15,10 +15,8 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
-use Illuminate\Support\Arr;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Auth;
 
 /**
  * Class Report.
@@ -140,26 +138,6 @@ class Report extends Model implements Nameable, HasSubtitle
     public function scopePending(Builder $query): void
     {
         $query->where(Report::ATTRIBUTE_STATUS, ApprovableStatus::PENDING->value);
-    }
-
-    /**
-     * Create a report with the given steps.
-     *
-     * @param  ReportStep|ReportStep[]  $steps
-     * @param  string|null  $notes
-     * @return static
-     */
-    public static function makeReport(ReportStep|array $steps, ?string $notes = null): static
-    {
-        $report = Report::query()->create([
-            Report::ATTRIBUTE_USER => Auth::id(),
-            Report::ATTRIBUTE_STATUS => ApprovableStatus::PENDING->value,
-            Report::ATTRIBUTE_NOTES => $notes,
-        ]);
-
-        $report->steps()->saveMany(Arr::wrap($steps));
-
-        return $report;
     }
 
     /**
