+++ app/Models/Admin/Report/ReportStep.php
@@ -12,8 +12,6 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
-use Illuminate\Database\Eloquent\Relations\Pivot;
-use Illuminate\Support\Arr;
 use Illuminate\Support\Carbon;
 
 /**
@@ -167,93 +165,6 @@ class ReportStep extends Model
         }
 
         return $newFields;
-    }
-
-    /**
-     * Create a report step to create a model.
-     *
-     * @param  class-string<Model>  $model
-     * @param  array  $fields
-     * @return self
-     */
-    public static function makeForCreate(string $model, array $fields): self
-    {
-        return static::makeFor(ReportActionType::CREATE, $model, $fields);
-    }
-
-    /**
-     * Create a report step to edit a model.
-     *
-     * @param  Model  $model
-     * @param  array  $fields
-     * @return self
-     */
-    public static function makeForUpdate(Model $model, array $fields): self
-    {
-        return static::makeFor(ReportActionType::UPDATE, $model, $fields);
-    }
-
-    /**
-     * Create a report step to delete a model.
-     *
-     * @param  Model  $model
-     * @return self
-     */
-    public static function makeForDelete(Model $model): self
-    {
-        return static::makeFor(ReportActionType::DELETE, $model);
-    }
-
-    /**
-     * Create a report step to attach a model to another in a many-to-many relationship.
-     *
-     * @param  Model  $foreign
-     * @param  Model  $related
-     * @param  class-string<Pivot>  $pivot
-     * @param  array  $fields
-     * @return self
-     */
-    public static function makeForAttach(Model $foreign, Model $related, string $pivot, array $fields): self
-    {
-        return static::makeFor(ReportActionType::ATTACH, $foreign, $fields, $related, $pivot);
-    }
-
-    /**
-     * Create a report step to detach a model from another in a many-to-many relationship.
-     *
-     * @param  Model  $foreign
-     * @param  Model  $related
-     * @param  Pivot  $pivot
-     * @param  array  $fields
-     * @return self
-     */
-    public static function makeForDetach(Model $foreign, Model $related, Pivot $pivot, array $fields): self
-    {
-        return static::makeFor(ReportActionType::DETACH, $foreign, $fields, $related, $pivot);
-    }
-
-    /**
-     * Create a report step for given action.
-     *
-     * @param  ReportActionType  $action
-     * @param  class-string<Model>|Model  $model
-     * @param  array|null  $fields
-     * @param  Model|null  $related
-     * @param  Pivot|null  $pivot
-     * @return self
-     */
-    protected static function makeFor(ReportActionType $action, Model|string $model, ?array $fields = null, ?Model $related = null, Pivot|string|null $pivot = null): self
-    {
-        return new self([
-            ReportStep::ATTRIBUTE_ACTION => $action->value,
-            ReportStep::ATTRIBUTE_ACTIONABLE_TYPE => $model instanceof Model ? $model->getMorphClass() : $model,
-            ReportStep::ATTRIBUTE_ACTIONABLE_ID => $model instanceof Model ? $model->getKey() : null,
-            ReportStep::ATTRIBUTE_FIELDS => Arr::where($fields, fn ($value, $key) => $model->isFillable($key)),
-            ReportStep::ATTRIBUTE_STATUS => ApprovableStatus::PENDING->value,
-            ReportStep::ATTRIBUTE_TARGET_TYPE => $related instanceof Model ? $related->getMorphClass() : null,
-            ReportStep::ATTRIBUTE_TARGET_ID => $related instanceof Model ? $related->getKey() : null,
-            ReportStep::ATTRIBUTE_PIVOT_CLASS => $pivot instanceof Model ? $pivot->getMorphClass() : $pivot,
-        ]);
     }
 
     /**
